[HOTFIX] EXIT을 DELETE에 작성했던 것을 수정함
