+++ src/main/java/com/hgcastle/common/ExitTemplate.java
@@ -1,5 +1,6 @@
 package com.hgcastle.common;
 
+import com.hgcastle.mapper.MemberExitMapper;
 import com.hgcastle.mapper.MemberMapper;
 import org.apache.ibatis.datasource.pooled.PooledDataSource;
 import org.apache.ibatis.mapping.Environment;
@@ -24,7 +25,7 @@ public class ExitTemplate {
                     new JdbcTransactionFactory(),
                     new PooledDataSource(DRIVER, URL, USER, PASSWORD));
             Configuration configuration = new Configuration(environment);
-            configuration.addMapper(MemberMapper.class);
+            configuration.addMapper(MemberExitMapper.class);
             sqlSessionFactory = new SqlSessionFactoryBuilder().build(configuration);
         }
         return sqlSessionFactory.openSession(false);
