+++ src/main/java/com/hgcastle/common/Template.java
@@ -1,6 +1,7 @@
 package com.hgcastle.common;
 
 import com.hgcastle.mapper.MemberDeleteMapper;
+import com.hgcastle.mapper.MemberMapper;
 import org.apache.ibatis.datasource.pooled.PooledDataSource;
 import org.apache.ibatis.mapping.Environment;
 import org.apache.ibatis.session.Configuration;
@@ -24,7 +25,7 @@ public class Template {
                     new JdbcTransactionFactory(),
                     new PooledDataSource(DRIVER, URL, USER, PASSWORD));
             Configuration configuration = new Configuration(environment);
-            configuration.addMapper(MemberDeleteMapper.class);
+            configuration.addMapper(MemberMapper.class);
             sqlSessionFactory = new SqlSessionFactoryBuilder().build(configuration);
         }
         return sqlSessionFactory.openSession(false);
