+++ src/main/java/com/hgcastle/controller/ExitController.java
@@ -1,35 +1,35 @@
 package com.hgcastle.controller;
 
 
+
 import com.hgcastle.dto.MemberDTO;
 import com.hgcastle.service.ExitService;
+import com.hgcastle.view.ExitResult;
 
 import java.util.Map;
 
 public class ExitController {
 
-//    private final PrintResult printResult;
-//    private final ExitService exitService;
-//
-//
-//    public ExitController() {
-//        printResult = new PrintResult();
-//        exitService = new ExitService();
-//
-//    }
-//
-//    public void selectId(Map<String, String> parameter){
-//
-//        String id = parameter.get("id");
-//
-//        MemberDTO member = exitService.selectId("id");
-//
-//        if (id != null){
-//            printResult.printMember(member);
-//        } else {
-//            printResult.printErrorMessage("selectOne");
-//        }
-//
-//    }
+    private final ExitService exitService;
+    private final ExitResult exitResult;
+
+    public ExitController () {
+
+        exitService = new ExitService();
+        exitResult = new ExitResult();
+    }
+
+    public void exitMembersById(Map<String, String> parameter) {
+
+        String id = parameter.get("id");
+        MemberDTO memberDTO = new MemberDTO();
+        memberDTO.setId(id);
+        if(exitService.exitMembersById(memberDTO)) {
+            exitResult.printSuccessMessage("exitMemberSuccess");
+        } else {
+            exitResult.printErrorMessage("exitMemberFail");
+        }
+
+    }
 
 }
