+++ src/main/java/com/hgcastle/mapper/MemberDeleteMapper.java
@@ -3,5 +3,5 @@ package com.hgcastle.mapper;
 import com.hgcastle.dto.MemberDTO;
 
 public interface MemberDeleteMapper {
-    int deleteMembersById(MemberDTO memberDTO);
+
 }
