+++ src/main/java/com/hgcastle/mapper/MemberExitMapper.java
@@ -4,5 +4,5 @@ import com.hgcastle.dto.MemberDTO;
 
 public interface MemberExitMapper {
 
-    MemberDTO selectId(String id);
+    int exitMembersById(MemberDTO memberDTO);
 }
