+++ src/main/java/com/hgcastle/service/ExitService.java
@@ -2,30 +2,34 @@ package com.hgcastle.service;
 
 
 import com.hgcastle.dto.MemberDTO;
+import com.hgcastle.mapper.MemberExitMapper;
 import org.apache.ibatis.session.SqlSession;
 
+import static com.hgcastle.common.ExitTemplate.getSqlSession;
 
-import static com.hgcastle.common.Template.getSqlSession;
 
 public class ExitService {
 
-//    private final ExitDAO exitDAO;
-//
-//    public ExitService(){
-//        exitDAO = new ExitDAO();
-//
-//    }
-//
-//    public MemberDTO selectId(String id) {
-//
-//        SqlSession sqlSession = getSqlSession();
-//
-//        MemberDTO member = exitDAO.selectId(sqlSession,id);
-//
-//        sqlSession.close();
-//
-//        return member;
-//    }
+    private MemberExitMapper mapper;
+
+    public boolean exitMembersById(MemberDTO memberDTO) {
+
+        SqlSession sqlSession = getSqlSession();
+
+        mapper = sqlSession.getMapper(MemberExitMapper.class);
+
+        int result = mapper.exitMembersById(memberDTO);
+
+        if (result > 0) {
+            sqlSession.commit();
+        } else {
+            sqlSession.rollback();
+        }
+
+        sqlSession.close();
+
+        return result > 0 ? true : false;
+    }
 
 }
