+++ src/main/resources/com/hgcastle/mapper/MemberDeleteMapper.xml
@@ -2,13 +2,6 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.hgcastle.mapper.MemberDeleteMapper">
-    <update id="deleteMembersById" parameterType="com.hgcastle.dto.MemberDTO">
-        update tbl_member
-        set
-        member_quitDate = curDate()
-        where
-        member_id = #{ id }
-        and member_quitdate is not null
-    </update>
+<mapper namespace="com.hgcastle.mapper.MemberMapper">
+
 </mapper>
